refactor: give both remote-fetch guards one address policy
Three review rounds each found this branch's hand-rolled BlockList copy of
"is this address safe" disagreeing with the audited one in safeRemoteFetch,
in both directions and in five places: it missed the 6to4 relay anycast
range and the RFC 6666 discard prefix, and it blanket-blocked the NAT64 and
IPv4-compatible ranges where the original decodes the IPv4 they carry and
judges that. The last of those was an availability bug, not a hole — on a
DNS64/NAT64 deployment every public IPv4-only origin resolves inside
64:ff9b::/96, so every Strava photo and remote attachment would have been
refused while ActivityPub traffic kept working.

Rather than add the two missing ranges and wait for round four to find a
sixth, the policy moves to lib/utils/unsafeAddress.ts and both guards
import it. The move is verbatim, and safeRemoteFetch's own tests pass
unchanged. A differential test over 44 address forms now pins that the two
answer identically, so the copy cannot silently drift back.

Separately, the protocol-relative check could not do its job as a raw
string prefix test. The WHATWG parser reads a backslash as a slash for a
special scheme and strips tab, LF and CR from the input before parsing, so
/\evil.example/... and /<TAB>/evil.example/... both carry their own
authority without beginning with //. All six variants were read as local
storage paths. It now decides on the RESOLVED authority instead.
